[BOOKINGSG-9317][GZ] separate local eslint rules to its own file
